+++ src/pages/NotebooksDashboardPage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { TrpcProvider } from "@/components/TrpcProvider.tsx";
-import { NotebookDashboard } from "@/components/notebooks/NotebookDashboard.tsx";
+import { NotebookDashboard } from "@/components/notebooks/dashboard/NotebookDashboard";
 
 export const NotebooksDashboardPage: React.FC = () => {
   return (
